requirements.txt
pdfplumber
pandas
python-docx
fuzzywuzzy
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
